Fix issue when Podfile without "use_frameworks!"
Using Podfile without "use_frameworks!" can generate issues.
In that case `#import <SnowplowTracker/SnowplowTracker-Swift.h>` raises the error "file not found".
Indeed the file doesn't exist and we have to use `#import "SnowplowTracker-Swift.h"`.
This is a temporary workaround in order to build the tracker with and without "use_frameworks!" directive on Podfile.